Fixed domains without expiration date list
It seems I used a list 2 years ago which was compiled for domains without a grace period, which is not the same as domains without expiration date. I am very sorry for that.

This list might still be incomplete. I checked all these domains manually and removed those which have an expiration date. I also added "eu".